fix(api): BindingRow.channel_status 改为 i32,对齐 channel_configs.status INT4 与全库约定,修复绑定列表 500

+++ backend/src/api/upstream_asset_bindings.rs
@@ -38,7 +38,7 @@ pub struct BindingRow {
     pub updated_at: Option<String>,
     pub channel_name: Option<String>,
     pub channel_base_url: Option<String>,
-    pub channel_status: Option<i64>,
+    pub channel_status: Option<i32>,
 }
 
 #[derive(Deserialize)]
